+++ addon/globalPlugins/contactsManager/__init__.py
@@ -18,8 +18,8 @@ import globalPluginHandler
 import globalVars
 import gui
 import wx
-from logHandler import log
 from scriptHandler import script
+from logHandler import log
 
 from .configPanel import ContactsManagerSettingsPanel
 from .main import ContactList