+++ addon/globalPlugins/contactsManager/lib/masked/ipaddrctrl.py
@@ -23,7 +23,7 @@ user hits '.' when typing.
 """
 
 import  wx
-from masked import BaseMaskedTextCtrl
+from wx.lib.masked import BaseMaskedTextCtrl
 
 # jmg 12/9/03 - when we cut ties with Py 2.2 and earlier, this would
 # be a good place to implement the 2.3 logger class